Add files via upload
Wrote code for adding points to player scores when the get a question right, added code for disabling a question permanently after a player answers it correctly, added code for showing a button when all questions have been answered. Added questions and answers for the history category. "answeredquestions.txt" should be empty at first, but I could not add it to github without something in it, so I added empty quotes, you can remove the empty quotes if you want. "remainingquestions.txt" should initially have the number of questions available at the beginning of the game.